Load in our own ts-node shim
