+++ src/config/ava-tsc.ts
@@ -3,7 +3,7 @@ import { resolve } from 'path'
 
 import { SRC_PATH, TESTS_NAME, DIST_PATH } from '../shared/constants'
 
-const TS_NODE_PATH = require.resolve('ts-node/register')
+const TS_NODE_PATH = require.resolve('../shim/ts-node')
 
 const getConfig = async () => {
   const files = await globby(TESTS_NAME, { cwd: SRC_PATH })
